Refactor resources & write filter for peer to peer unicast (#959)
* Refactor transport peers (#36)

* refactor: transport peers

* fix: set peer common as first member

* Refactor resource system (#37)

* feat: remove unused peer_id

* refactor: mapping is now peer address as id

* feat: store remote resource at peer level

* fix: missing fix mapping

* fix: missing init & memory leak

* feat: propagate peer in triggers

* Rework write filters (#38)

* feat: remove unused peer_id

* feat: store remote resource at peer level

* fix: missing fix mapping

* feat: propagate peer in triggers

* feat: switch write filter to decl id list

* feat: use peer addr to avoid id collisions

* feat: update filter on peer disconnection

* fix: add comment

* feat: avoid recursive list parsing

* feat: invalidate cache on peer disconnection

* feat: push declarations on new unicast peer

* feat: activate ke declaration on p2p unicast

* fix: remove filter state init

* feat: add replay declare function

* fix: deactivate write filter on multicast

* fix: read declare on single thread examples

---------

+++ include/zenoh-pico/protocol/definitions/declarations.h
@@ -110,7 +110,7 @@ typedef struct {
     } _body;
 } _z_declaration_t;
 void _z_declaration_clear(_z_declaration_t* decl);
-void _z_decl_fix_mapping(_z_declaration_t* msg, uint16_t mapping);
+void _z_decl_fix_mapping(_z_declaration_t* msg, uintptr_t mapping);
 
 _z_declaration_t _z_make_decl_keyexpr(uint16_t id, _Z_MOVE(_z_keyexpr_t) key);
 _z_declaration_t _z_make_undecl_keyexpr(uint16_t id);
