+++ src/protocol/definitions/network.c
@@ -360,7 +360,7 @@ _z_network_message_t *_z_n_msg_clone(const _z_network_message_t *src) {
     return dst;
 }
 
-void _z_msg_fix_mapping(_z_zenoh_message_t *msg, uint16_t mapping) {
+void _z_msg_fix_mapping(_z_zenoh_message_t *msg, uintptr_t mapping) {
     switch (msg->_tag) {
         case _Z_N_DECLARE: {
             _z_decl_fix_mapping(&msg->_body._declare._decl, mapping);
